チャット時の引数を変更

+++ Ollapi/api/OllapiChatResponse.cs
@@ -25,21 +25,21 @@ namespace Ollapi.api
         public bool Done { get; set; }
 
         [JsonPropertyName("total_duration")]
-        public int TotalDuration {  get; set; }
+        public Int64 TotalDuration {  get; set; }
         
         [JsonPropertyName("load_duration")]
-        public int LoadDuration {  get; set; }
+        public Int64 LoadDuration {  get; set; }
         
         [JsonPropertyName("prompt_eval_count")]
-        public int PromptEvalCount {  get; set; }
+        public Int64 PromptEvalCount {  get; set; }
         
         [JsonPropertyName("prompt_eval_duration")]
-        public int PromptEvalDuration {  get; set; }
+        public Int64 PromptEvalDuration {  get; set; }
         
         [JsonPropertyName("eval_count")]
-        public int EvalCount {  get; set; }
+        public Int64 EvalCount {  get; set; }
         
         [JsonPropertyName("eval_duration")]
-        public int EvalDuration { get; set; }
+        public Int64 EvalDuration { get; set; }
     }
 }
